+++ src/utils.cpp
@@ -1,0 +1,25 @@
+#include "utils.h"
+#include <random>
+#include <chrono>
+
+unsigned int utils::getTimestamp()
+{
+    return std::chrono::time_point_cast<std::chrono::microseconds>(
+            std::chrono::high_resolution_clock::now()).time_since_epoch().count();
+}
+unsigned short int utils::randomNum(const unsigned short int *maxNum)
+{
+    unsigned int seed = std::chrono::system_clock::now().time_since_epoch().count();
+    std::mt19937_64 generator(seed);
+    return generator() % *maxNum; // lowers the number to the specified range
+}
+utils::timer::timer(unsigned int millisecondsToWait)
+{
+    startTime = getTimestamp();
+    timeToWait = millisecondsToWait;
+}
+bool utils::timer::done() const
+{
+    unsigned int timeTaken = getTimestamp() - startTime;
+    return timeTaken / 1000 > timeToWait;
+}
